organizationManagement/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.14
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itsdangerous==0.24
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.3
mysqlclient==2.1.1
packaging==23.1
pika==1.3.0
pycodestyle==2.9.1
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
